refactor(compiler): organize array code and improve cfuncs structure
- Move array functions from compiler.go to array.go for better organization
- Reorder ARR_* constants in cfuncs.go by type (I64, F64, STR) then function
- Add spacing between different function types in cfuncs.go
- Organize GetFnType switch to match constant order with clear sections
- Maintain all existing functionality with cleaner code structure
